Thoroughly test LogContent
